refactored executor and workers into their own header files
