+++ src/include/utils.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <config.hpp>
+#include <filesystem>
 #include <fstream>
 #include <vector>
